apps/trialist/requirements.txt
pyyaml
flask
Flask-HTTPAuth
flask-restplus
gunicorn
markdown>=3.0
jinja2
python-dateutil